### 2021-08-30 #3670 @whedon
[PRE REVIEW]: GemGIS - Spatial Data Processing for Geomodeling
**Submitting author:** @AlexanderJuestel (<a href="http://orcid.org/0000-0003-0980-7479">Alexander Jüstel</a>)
**Repository:** <a href="https://github.com/cgre-aachen/gemgis" target ="_blank">https://github.com/cgre-aachen/gemgis</a>
**Version:** v1.0.0
