### 2021-08-30 #3670 @arfon
:wave: @AlexanderJuestel, thanks for your submission to JOSS. It looks like the `paper.md` needs a little work. In particular, every author needs an affiliation, and ORCIDs can't be blank (but you can just remove the line if an author doesn't have one).

You can check the paper compiles after making changes with `@whedon generate pdf`
