### 2021-08-30 #3670 @crvernon
:wave: @AlexanderJuestel - I'll be your topic editor for this submission.

The first step in this pre-review will be to see to the comment above about getting the paper in a format that we can compile.
 Next we can find at least 2 reviewers and get the review started. If you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, this list of people have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).  Thanks!
