### 2021-08-30 #3670 @crvernon
Also, I noticed that `Miguel de la Varga` has made commits to your repository.  Is there a reason that they are not included in the author list?
