### 2021-08-30 #3670 @AlexanderJuestel
@crvernon, thanks for being the topic editor. The changes suggested by @arfon were implemented. The article is compiling now. I also added @leguark (Miguel de la Varga) as author, thanks for pointing that out.

As reviewers, I would like to suggest **martinfleis** as he is the author of the GeoPandas package, which is extensively used in GemGIS, and **banesullivan** as we use his PyVista package as well.
