### 2021-09-01 #3670 @omshinde
Hi @crvernon
Thanks for the mention. I would be happy to review the submission on "GemGIS - Spatial Data Processing for Geomodeling". Kindly assign me as a reviewer. Regards.
