### 2021-09-09 #3670 @martibosch
Hello @crvernon, thank you for considering me but I am afraid that I will have to decline your kind invitation since I am also very busy at the moment and I further have another JOSS review pending. Best,

Martí
