### 2021-09-15 #3709 @crvernon
Thanks for checking in @AlexanderJuestel - at this point we will wait on the first round of feedback to come in as issues/comments. You can address these as they come up to expedite the process. I'll periodically send out check-in comments that open up lines for us to communicate any blockers to progress.
