### 2021-09-16 #3709 @kanishkan91
@AlexanderJuestel I've started the review. I had some very minor questions for you related to the repository,
1. Firstly, there are two pull requests that are open [here](https://github.com/cgre-aachen/gemgis/pulls). Do these need to be merged in? Just wanted to confirm.
2. I looked at the recent commit history. It looked like the recent commits were mostly from yourself and @RichardScottOZ. Could you describe the roles that the other co-authors played in the development of the software repository?
3. Also, in the pydocs, the main developers from the pydocs file - https://gemgis.readthedocs.io/en/latest/getting_started/authors.html don't all appear as co-authors in the paper. Is this intentional? This being a software paper, ideally all major contributors to software development that are mentioned in the docs should be co-authors.
4.  The license you use is GNU LESSER GENERAL PUBLIC LICENSE. My understanding is limited here, but a GPL license would impose some copy left restrictions on the users (they would have to ensure that they release anything that uses GemGIS as a freely available open source software package). Is this intentional? I have seen packages use BSD2 in its place which offers more flexibility to the end user. @crvernon Do you have any thoughts on this?
