### 2021-09-17 #3709 @AlexanderJuestel
@kanishkan91, thanks for getting started on the review. Regarding your comments:

1. I merged one of the PRs but would like to keep the second one open as we may change some of the tutorial notebooks during the review process here in combination with updating the documentation.

2. @RichardScottOZ is a very active user providing always great feedback on typos and grammatical errors in the docstrings or notebooks. However, he has not played an active role in the development of the software yet.

3. Good point. As for @RichardScottOz, @Japhiolite provided great feedback on the code and the documentation, mostly offline,  but did not provide any functionality for the software. I would add 'Marius Pischke' and 'Miguel de la Varga' as authors to the documentation and change the status of @RichardsScottOz and @Japhiolite.
