### 2021-09-17 #3709 @AlexanderJuestel
@kanishkan91 I just fixed the authors in the docs. The current build of the documentation is failing though, so changes are only visible locally right now but let's tackle one thing at a time :)
