### 2021-09-17 #3709 @kanishkan91
@AlexanderJuestel  Thanks. For my point no 2 from above, could you describe the roles the different co-authors played in the development of the software? This does not need to be a part of the paper or repo but having those descriptions here in the review thread would suffice.
