### 2021-09-17 #3709 @AlexanderJuestel
@kanishkan91, I am the main developer of the software and added most of the functionality of the package. @Endarthur contributed with helpful tips with respect to the implementation of certain functionality. He basically helped me to make things run. Marius Pischke developed with me functionality within the framework of his master thesis at RWTH Aachen University, which I then implemented in GemGIS as he does not have a Github account. @Leguark is the main developer for the GemPy package (https://github.com/cgre-aachen/gempy). Without his developments, GemGIS would not have a purpose. He also contributed minor things to GemGIS as well. @flohorovicic is/was my supervisor who highly encouraged me to develop GemGIS. @Japhiolite and @RichardScottOz mainly provided feedback on the documentation and notebooks and also provided feedback on other communication channels.

Would this description be sufficient or do you need any more information or statements of the respective persons?
