### 2021-09-22 #3709 @crvernon
:mega:  Mid-week rally!  Just checking in to see how things are going.

:open_hands: @kanishkan91 thanks for kicking off this review!

 @omshinde how are things going?  The author @AlexanderJuestel has been very responsive if you have any questions.

Have a great and productive rest of the week!
