### 2021-09-23 #3709 @omshinde
@crvernon Thanks for the reminder. I just finished my other JOSS review and would be able to progress fast with GemGIS submission now. I am very excited and looking forward.
