### 2021-10-16 cgre-aachen/gemgis#193 @kanishkan91
JOSS review - Consider adding testing framework using GITHUB actions in addition to binder
Currently the testing framework environment used by this repository is binder. I would recommend that this also be implemented for Github actions. This may make it more robust to any failures introduced by future development.

https://github.com/openjournals/joss-reviews/issues/3709
