### 2021-10-16 cgre-aachen/gemgis#194 @kanishkan91
JOSS review - Consider adding Coverage.py
I would recommend that the author add in a coverage.py file to this repo as described here - https://coverage.readthedocs.io/en/6.0.2/

This tells the user what the coverage of the tests is like. If the author does not want to share this info, that's fine, but this should be added to the repo as more users make use of gemgis

https://github.com/openjournals/joss-reviews/issues/3709
