### 2021-10-16 cgre-aachen/gemgis#195 @kanishkan91
JOSS review :  Consider adding dashboard using python dash
gemgis is clearly a powerful tool that creates data objects as would be required by GemPy. In particular I like how a user can not only make use of gemgis to create a data object but can also make use of this software to do meaningful post processing of GemPy outputs. But, these functionalities can be very effectively illustrated using a dashboard (I would recommend Dash -https://dash.plotly.com/).

In particular Example 1 can be presented using a dashboard as opposed to the text heavy example as it stands right now. As the repo currently stands it is very much usable but not entirely digestable for novice users (the ones who would most benefit from something like this given that they would find GemPy intimidating.) Having that dashboard would effectively illustrate the utility of this software (which is very high). But more importantly, it would help a novice user understand the functionality much better and would help them effectively use this software.

https://github.com/openjournals/joss-reviews/issues/3709
