### 2021-10-16 #3709 @kanishkan91
@AlexanderJuestel  I am going through this review and am adding comments as I go through. I hope to be done by Sunday night. But there is one particular feature that I think would highlight the utility of what you have built much more. Also considering the functionality already built, I think this would be low hanging fruit. I have created an issue already but I add the text from the issue below since I think we should start a discussion on this sooner rather than later.

gemgis is clearly a powerful tool that creates data objects as would be required by GemPy. In particular I like how a user can not only make use of gemgis to create a data object but can also make use of this software to do meaningful post processing of GemPy outputs. But, these functionalities can be very effectively illustrated using a dashboard (I would recommend Dash -https://dash.plotly.com/).

In particular Example 1 can be presented using a dashboard as opposed to the text heavy example as it stands right now. As the repo currently stands it is very much usable but not entirely digestible for novice users (the ones who would most benefit from something like this given that they would find GemPy intimidating.) Having that dashboard would effectively illustrate the utility of this software (which is very high). But more importantly, it would help a novice user understand the functionality much better and would help them effectively use this software.
