### 2021-10-16 cgre-aachen/gemgis#196 @kanishkan91
JOSS Review : Documentation updates
I request many updates to the documentation. In particular, while the documentation is dense, much of it is geared towards explaining GIS processing using other softwares. I suggest that the author split the documentation into two parts,

1. Documentation of GIS functionality in gemgis (This should include what functions are exclusively available in gemgis)
2. Introduction to basic GIS functionality in other software

Also, it would be better if the author opened with the example (which comes in the end right now). This would improve the documentation.

I did not understand the API reference. It seems to be incomplete? But maybe I missed something?

https://github.com/openjournals/joss-reviews/issues/3709
