### 2021-10-17 cgre-aachen/gemgis#197 @kanishkan91
JOSS review: Add explicit documentation of the GemPy data class  
The GemPy class that's created and used seems to be an important part of the utility of this software. I would recommend the addition of documentation explicitly walking the user through what the different attributes of this class are. This can be added as a table in the readme. Having this documentation would be useful for future development as well. For example, as the GemPy  evolves,  there may be new attributes that get added or old ones removed. Having this documented is very useful for both developers and end users.

https://github.com/openjournals/joss-reviews/issues/3709
