### 2021-10-17 cgre-aachen/gemgis#198 @kanishkan91
JOSS review (readme modifications): Add contribution instructions , remove images etc.
There are currently no contribution instructions on the readme. Please add a section for the same. Also, please remove the images in the readme under the section "Gallery". These are not necessary since these concepts have been explained in detail in the documentation.

Also, please consider adding a flowchart to the readme (similar to one in the paper) which would illustrate that gemgis is not just a data processing software. It bridges the gap between GIS data processing and Geo modelling and thus creates an effective GIS processing environment.

https://github.com/openjournals/joss-reviews/issues/3709
