### 2021-10-17 cgre-aachen/gemgis#199 @kanishkan91
JOSS review: Explicit link to "vignette"
Instead of having an Example 1 in the documentation, the author can just create a vignette (i.e. a jupyter notebook similar to example 1) and add an explicit link to the same in the readme. This would make it easy for a new user to find as opposed to looking around in the documentation. This is an easy fix as the author can simply change the example to a vignette with some documentation edits and add an explicit link to the same in the readme.

https://github.com/openjournals/joss-reviews/issues/3709
