### 2021-10-20 #3709 @crvernon
:mega: Mid-week rally!  Looks like @AlexanderJuestel and @kanishkan91 have a good dialogue going and are keeping the review progressing.

:wave: @omshinde - Just checking in to see how things are going?  Let me know if you have any questions.

❗ @kanishkan91 and @omshinde - Don't forget to keep you checklist up-to-date as you work through the review.  Thanks!

:clap: Keep up the good work!
