### 2021-11-03 #3709 @crvernon
:mega: Mid-week rally!  @AlexanderJuestel @kanishkan91 and @omshinde could you provide a short update as to how things are going?   Looks like you are making good progress.  Let's keep driving towards getting things in shape for publication!

:clap: Keep up the good work!
