### 2021-11-17 #3709 @crvernon
:mega: Mid-week rally!

@kanishkan91 and @omshinde please provide a short update in this issue thread concerning the status of your review.  Is there anything you are waiting on?  Thanks!

@AlexanderJuestel how are things going on your side of things?  Could you provide a short update here as well?  I would like to keep this moving while we have good momentum. Thanks!
