### 2021-11-18 #3709 @AlexanderJuestel
@crvernon, thanks for checking in! Currently, I am a little limited in time due to my PhD studies. In addition, we are currently working on a new GemPy release so I wanted to wait for that :)
