### 2021-12-03 #3709 @omshinde
Hi @crvernon ! Thanks for the reminder.
I have submitted my initial review (https://github.com/openjournals/joss-reviews/issues/3709#issuecomment-953245623) and waiting for the authors response. Once they are addressed, I would be happy to proceed with the review.
