### 2021-12-03 #3709 @kanishkan91
@crvernon Apologies for the late message. I am also waiting on the author to respond to comments.

@AlexanderJuestel Let me know whenever and I can take a look again (I know its a busy time of year, so no hurry!). By the way its interesting that you mentioned that GemPy is also coming out with a release. This would be an interesting opportunity to see the documentation of the GemPy object documentation in gemgis when its ready.
