### 2022-01-07 cgre-aachen/gemgis#193 @AlexanderJuestel
@kanishkan91
There is a Github Action workflow for automated testing here: https://github.com/cgre-aachen/gemgis/blob/main/.github/workflows/testing.yml and the tests are located here: https://github.com/cgre-aachen/gemgis/tree/main/tests

Binder is not used for testing.
