### 2022-01-07 cgre-aachen/gemgis#195 @AlexanderJuestel
@kanishkan91 I would rather not include a dashboard to actually not overwhelm the user. GemPy is not using any fancy representation package except PyVista for 3D models so keeping it simple with Matplotlib plots or PyVista Plots would be the plan. What do you think?
