+++ README.md
@@ -58,943 +58,3 @@ We welcome issue reports, questions, ideas for new features and pull-requests to
 * de la Varga, M., Schaaf, A., and Wellmann, F.: GemPy 1.0: open-source stochastic geological modeling and inversion, Geosci. Model Dev., 12, 1-32, https://doi.org/10.5194/gmd-12-1-2019, 2019
 * Powell, D.: Interpretation of Geological Structures Through Maps: An Introductory Practical Manual, Longman, pp. 192, 1992
 * Bennison, G.M.: An Introduction to Geological Structures and Maps, Hodder Education Publication, pp. 78, 1990
-
-<a name="gallery"></a>
-## Gallery
-
-### Working with Vector Data
-
-<p>
-<table>
-
-<tr>
-    <td><b style="font-size:30px">Extracting XY values from Vector Data</b></td>
-    <td><b style="font-size:30px">Extracting XYZ values from Vector Data</b></td>
- </tr>
-<tr>
-<tr>
-
-<td>
-<a href="https://gemgis.readthedocs.io/en/latest/getting_started/tutorial/01_extract_xy.html">
-<img alt="extracting vertices from vector data" src="https://raw.githubusercontent.com/cgre-aachen/gemgis/main/docs/getting_started/images/tutorial01_cover.png" width="400"/>
-</a>
-</td>
-
-<td>
-<a href="https://gemgis.readthedocs.io/en/latest/getting_started/tutorial/02_extract_xyz.html">
-<img alt="extracting vertices from vector data" src="https://raw.githubusercontent.com/cgre-aachen/gemgis/main/docs/getting_started/images/tutorial02_cover.png" width="400"/>
-</a>
-</td>
-
-</tr>
-</table>
-</p>
-
-
-
-  
-
-<p>
-<table>
-
-<tr>
-    <td><b style="font-size:30px">Exploding Geometries/Vector Data</b></td>
-    <td><b style="font-size:30px">Clipping/Cropping Vector Data</b></td>
- </tr>
-<tr>
-
-<td>
-
-<a href="https://gemgis.readthedocs.io/en/latest/getting_started/tutorial/03_exploding_geometries.html">
-<img alt="exploding geometries" src="https://raw.githubusercontent.com/cgre-aachen/gemgis/main/docs/getting_started/images/tutorial03_cover.png" width="400" class="center"/>
-</a>
-</td>
-
-<td>
-<a href="https://gemgis.readthedocs.io/en/latest/getting_started/tutorial/04_clipping_data.html">
-<img alt="clipping vector dataa" src="https://raw.githubusercontent.com/cgre-aachen/gemgis/main/docs/getting_started/images/tutorial04_cover.png" width="400" class="center" />
-</a>
-</td>
-</tr>
-</table>
-</p>
-
-<p>
-<table>
-
-<tr>
-    <td><b style="font-size:30px">Interpolating Vector Data to Rasters</b></td>
-    <td><b style="font-size:30px">Removing Interface Points within Fault Buffers</b></td>
- </tr>
-<tr>
-
-<td>
-<a href="https://gemgis.readthedocs.io/en/latest/getting_started/tutorial/05_interpolating_rasters.html">
-<img alt="interpolating vector data" src="https://raw.githubusercontent.com/cgre-aachen/gemgis/main/docs/getting_started/images/tutorial05_cover.png" width="400" class="center" />
-</a>
-</td>
-
-<td>
-<a href="https://gemgis.readthedocs.io/en/latest/getting_started/tutorial/11_removing_interfaces_within_fault_buffers.html">
[…]
