### 2022-01-07 cgre-aachen/gemgis#198 @AlexanderJuestel
Hello @kanishkan91
- Contribution instructions were added
- Images were removed
- Flow Chart was added

Please see https://github.com/cgre-aachen/gemgis/pull/202
