### 2022-01-07 cgre-aachen/gemgis#203 @AlexanderJuestel
JOSS Review: Fixing GemGIS Joss Paper
From https://github.com/openjournals/joss-reviews/issues/3709#issuecomment-953245623

**Software paper**
The software paper seems to be short and here, I agree to the comments of the fellow reviewer that there is a scope of adding more details to the software paper (maybe adding one of the example tutorials as a case study of the proposed package). But this should be the authors judgement to extend the paper or not, as there is no such limit/rule.

I have following suggestions for this section:

The section titled Statement of need is missing in the software paper. I would suggest to take highlighting points from the GemGIS Target Audience and Background section and add it as bullet points in the Statement of need to give a crisp understanding to the readers.

- [x] The heading References is missing from the software paper and after Acknowledgments, directly the list of referenced articles are mentioned. Please correct me if I am missing anything here.
- [x] There is a typo in the paper at line 78, GemGIS Outlook. graviational -> gravitational

It is recommended to comprehensively check the paper for typos.
