### 2022-01-07 cgre-aachen/gemgis#204 @AlexanderJuestel
JOSS Review: Fixing documentation 
From https://github.com/openjournals/joss-reviews/issues/3709#issuecomment-953245623, @omshinde

The examples are very well defined and comprehensive for new users interested in using GemGIS package. I have following queries/suggestions associated with this section -

1. Functionality documentation: The links to the API references are not working or only headings are mentioned - (GemGIS Data Object or Working with vector data). Please correct me if I am missing something here.

2. Automated tests: Although there is a section describing the process to check the installation in the documentation, but automated tests are not defined. https://gemgis.readthedocs.io/en/latest/getting_started/installation.html#checking-the-installation. Meanwhile I could find the description for automated tests on the project Readme so I would suggest authors to replicate this page on the documentation too.

3. Community guidelines: I could not find community guidelines in the documentation. But similar to tests, the authors could replicate the guidelines from the Readme as mentioned here.
