+++ docs/index.rst
@@ -14,6 +14,7 @@ Welcome to documentation page of GemGIS - GemPy Geographic!
    getting_started/about
    getting_started/authors
    getting_started/installation
+   getting_started/contributing
    getting_started/data_types
    getting_started/whatiswhat/index
    getting_started/tutorial/index
