### 2022-01-07 cgre-aachen/gemgis#204 @AlexanderJuestel
See https://github.com/cgre-aachen/gemgis/pull/202/files

Automated tests and contribution guidelines were added to the documentation. Now I just need to get the documentation running again
