+++ environment_dev.yml
@@ -11,6 +11,7 @@ dependencies:
   - pip:
     - gemgis
     - matplotlib
+    - pandas
     - pooch<1.5.0
    # - gempy==2.2.9
     - pyqt5
