### 2022-01-07 cgre-aachen/gemgis#195 @kanishkan91
Yes. This sounds good for the purpose of the paper (It will be a bit of overkill).
