+++ environment_dev.yml
@@ -4,7 +4,7 @@ channels:
 dependencies:
   #- python>=3.7
   #- pygeos=0.9
-  #- geopandas=0.9
+  - geopandas
   - rasterio
   #- pyvista=0.29
   - pip
