+++ README_old.md
@@ -19,9 +19,8 @@ We attempt to simplify the access to open-source spatial data processing for geo
 
 GemGIS wraps and extends the functionality of packages known to the geo-community such as [GeoPandas](https://geopandas.org/), [rasterio](https://rasterio.readthedocs.io/en/latest/#), [OWSLib](https://geopython.github.io/OWSLib/), [Shapely](https://shapely.readthedocs.io/en/latest/manual.html), [PyGEOS](https://pygeos.readthedocs.io/en/latest/), [PyVista](https://docs.pyvista.org/), [Pandas](https://pandas.pydata.org/), [NumPy](https://numpy.org/) and the geomodeling package [GemPy](https://docs.gempy.org/). 
 
-The aim of GemGIS, as indicated by the name, is to become a bridge between conventional geoinformation systems (GIS) such as ArcGIS and QGIS, and geomodeling tools such as GemPy, allowing simpler and more automated workflows from one environment to the other. This also includes making it simpler to visualize the results obtained from GemGIS and GemPy with PyVista or Blender. 
+The aim of GemGIS, as indicated by the name, is to become a bridge between conventional geoinformation systems (GIS) such as ArcGIS and QGIS, and geomodeling tools such as GemPy, allowing simpler and more automated workflows from one environment to the other.
 
-<p align="center"><img src="https://raw.githubusercontent.com/cgre-aachen/gemgis/main/joss/images/fig1.png" width="800">
 
 <a name="doc"></a>
 ## Resources
@@ -59,3 +58,942 @@ We welcome issue reports, questions, ideas for new features and pull-requests to
 * Powell, D.: Interpretation of Geological Structures Through Maps: An Introductory Practical Manual, Longman, pp. 192, 1992
 * Bennison, G.M.: An Introduction to Geological Structures and Maps, Hodder Education Publication, pp. 78, 1990
 
+<a name="gallery"></a>
+## Gallery
+
+### Working with Vector Data
+
+<p>
+<table>
+
+<tr>
+    <td><b style="font-size:30px">Extracting XY values from Vector Data</b></td>
+    <td><b style="font-size:30px">Extracting XYZ values from Vector Data</b></td>
+ </tr>
+<tr>
+<tr>
+
+<td>
+<a href="https://gemgis.readthedocs.io/en/latest/getting_started/tutorial/01_extract_xy.html">
+<img alt="extracting vertices from vector data" src="https://raw.githubusercontent.com/cgre-aachen/gemgis/main/docs/getting_started/images/tutorial01_cover.png" width="400"/>
+</a>
+</td>
+
+<td>
+<a href="https://gemgis.readthedocs.io/en/latest/getting_started/tutorial/02_extract_xyz.html">
+<img alt="extracting vertices from vector data" src="https://raw.githubusercontent.com/cgre-aachen/gemgis/main/docs/getting_started/images/tutorial02_cover.png" width="400"/>
+</a>
+</td>
+
+</tr>
+</table>
+</p>
+
+
+
+  
+
+<p>
+<table>
+
+<tr>
+    <td><b style="font-size:30px">Exploding Geometries/Vector Data</b></td>
+    <td><b style="font-size:30px">Clipping/Cropping Vector Data</b></td>
+ </tr>
+<tr>
+
+<td>
+
+<a href="https://gemgis.readthedocs.io/en/latest/getting_started/tutorial/03_exploding_geometries.html">
[…]
