+++ environment_dev.yml
@@ -4,9 +4,9 @@ channels:
 dependencies:
   #- python>=3.7
   #- pygeos=0.9
-  - geopandas
-  - rasterio
-  - pyvista
+  #- geopandas
+  #- rasterio
+  #- pyvista
   - pip
   - pip:
     - gemgis
