+++ .readthedocs.yaml
@@ -18,7 +18,7 @@ formats:
 
 # Optionally set the version of Python and requirements required to build your docs
 python:
-   version: 3.9
+   version: 3.8
 
 conda:
    environment: environment_dev.yml
