### 2022-01-10 cgre-aachen/gemgis#199 @kanishkan91
@AlexanderJuestel - Your main vignette (walk through example) is here - https://gemgis.readthedocs.io/en/latest/getting_started/examples/example01_parallel_inclined_layers.html. You can directly add a link to this from your readme. You can add a section called "vignette" or "example with walkthrough of gemgis" and just add a hyperlink to the example. Does this make sense?
