+++ docs/index.rst
@@ -21,19 +21,6 @@ Welcome to documentation page of GemGIS - GemPy Geographic!
    getting_started/example_models/index
 
 
-..
-   .. toctree::
-      :maxdepth: 2
-      :caption: API Reference
-      api_reference/overview
-      api_reference/gemgis_object
-      api_reference/vector_data
-      api_reference/raster_data
-      api_reference/online_services
-      api_reference/utility_tool
-      api_reference/visualization
-      api_reference/misc
-      api_reference/postprocessing
 
 
 
