+++ docs/api_reference/raster_data.rst
@@ -1,67 +1,0 @@
-.. _raster_data_ref:
-
-Working with Raster Data
-===========================================================
-
-..automodule:: gemgis.raster
-
-Sampling from Raster Data
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. autofunction:: gemgis.raster.sample_from_array
-
-.. autofunction:: gemgis.raster.sample_from_rasterio
-
-.. autofunction:: gemgis.raster.sample_randomly
-
-.. autofunction:: gemgis.raster.sample_orientations
-
-.. autofunction:: gemgis.raster.sample_interfaces
-
-Calculating Raster Properties
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. autofunction:: gemgis.raster.calculate_hillshades
-
-.. autofunction:: gemgis.raster.calculate_slope
-
-.. autofunction:: gemgis.raster.calculate_aspect
-
-.. autofunction:: gemgis.raster.calculate_difference
-
-Clipping Raster Data
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. autofunction:: gemgis.raster.clip_by_bbox
-
-.. autofunction:: gemgis.raster.clip_by_polygon
-
-Resizing Raster Data
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. autofunction:: gemgis.raster.resize_by_array
-
-.. autofunction:: gemgis.raster.resize_raster
-
-
-Reading different types of Raster/Mesh Data
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. autofunction:: gemgis.raster.read_msh
-
-.. autofunction:: gemgis.raster.read_ts
-
-.. autofunction:: gemgis.raster.read_asc
-
-.. autofunction:: gemgis.raster.read_zmap
-
-Opening and saving Raster Data
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. autofunction:: gemgis.raster.save_as_tiff
-
-.. autofunction:: gemgis.raster.create_filepaths
-
-.. autofunction:: gemgis.raster.create_src_list
-
-.. autofunction:: gemgis.raster.merge_tiles
