+++ docs/api_reference/postprocessing.rst
@@ -1,4 +1,0 @@
-.. _postprocessing_ref:
-
-Postprocessing of GemPy Models
-===========================================================
