+++ docs/api_reference/index.rst
@@ -1,21 +1,0 @@
-.. gemgis documentation master file, created by
-   sphinx-quickstart on Mon Nov  2 22:04:17 2020.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-Index
-===========================================================
-
-.. toctree::
-   :maxdepth: 2
-   :caption: API Reference
-
-   overview
-   gemgis_object
-   vector_data
-   raster_data
-   online_services
-   utility_tool
-   visualization
-   misc
-   postprocessing
