+++ docs/api_reference/overview.rst
@@ -1,9 +1,0 @@
-.. gemgis documentation master file, created by
-   sphinx-quickstart on Mon Nov  2 22:04:17 2020.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-Overview
-===========================================================
-
-Within the different fields of geosciences, different data types exist which are stored in many different data formats. The subsequent sections will elaborate on how to work with the designated ``GemGISData`` object, how to work with ``vector`` data, ``raster`` data and data obtained by ``online services``. The ``utils`` section provides utility functions to work with the previously described data types in more depth. The ``visualization`` and plotting section provides functions to simplify the plotting of data processed with GemGIS such as boreholes and cross sections. The ``misc`` section is dealing with specialized functions adapted for specific case studies. The ``postprocessing`` section provides functions to further use a created GemPy model.
