+++ docs/getting_started/index.rst
@@ -15,8 +15,7 @@ Index
    installation
    contributing
 
-   .. data_types
+
 
    whatiswhat/index
    tutorial/index
-   .. example_models/index
