+++ .readthedocs.yaml
@@ -9,7 +9,7 @@ version: 2
 sphinx:
    builder: html
    configuration: docs/conf.py
-   fail_on_warning: true
+   fail_on_warning: false
 
 # Optionally build your docs in additional formats such as PDF
 formats:
