+++ docs/conf.py
@@ -37,7 +37,7 @@ sys.path.append('../../gemgis')
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
-    #'nbsphinx',
+    'nbsphinx',
     'sphinx_rtd_theme',
     #'sphinx.ext.autodoc',
     'sphinx.ext.napoleon',
