+++ docs/getting_started/tutorial/index.rst
@@ -37,3 +37,4 @@ Each set of functions of GemGIS is collected in a different module. The function
     :caption: Tutorials
 
     00_generating_data_qgis
+    01_extract_xy
