### 2022-03-04 #3709 @AlexanderJuestel
@kanishkan91  I would like to ask you to check the following points off your checklist:

- **Example usage**: Here you can find many examples of using GemGIS: https://gemgis.readthedocs.io/en/latest/getting_started/tutorial/index.html

- **Functionality documentation**: I was just able to build the documentation again. See for example https://readthedocs.org/projects/gemgis/builds/16267667/

- **Community Guidelines**: Please see the contributing document for that: https://github.com/cgre-aachen/gemgis/blob/main/CONTRIBUTING.md
