+++ docs/conf.py
@@ -64,7 +64,6 @@ exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store', '**.ipynb_checkpoints']
 
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
-#
 html_theme = 'sphinx_rtd_theme'
 
 # Add any paths that contain custom static files (such as style sheets) here,
@@ -72,12 +71,12 @@ html_theme = 'sphinx_rtd_theme'
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = []
 
-#nbsphinx_execute = 'never'
+nbsphinx_execute = 'never'
 
-#nbsphinx_execute_arguments = [
-#    "--InlineBackend.figure_formats={'svg', 'pdf'}",
-#    "--InlineBackend.rc={'figure.dpi': 96}",
-#]
+nbsphinx_execute_arguments = [
+    "--InlineBackend.figure_formats={'svg', 'pdf'}",
+    "--InlineBackend.rc={'figure.dpi': 96}",
+]
 
 copybutton_prompt_text = ">>> "
 
