### 2022-03-08 #3709 @kanishkan91
@AlexanderJuestel I ticked off the items on my list. But I had a couple of follow up questions so that I can close all issues that I opened.

- Since the tests are all set up now, could you create a codecov badge for your repository , so I can check the the percentage of the code covered by tests? If you've not linked this repo to codecov, it is possible here -https://about.codecov.io/blog/python-code-coverage-using-github-actions-and-codecov/ . With this, you can ignore my comment on the Coverage.py.
- Could you take a look at my comment about cleaning up the readme? I still think you can take out some of the images (where some of the links are broken) and maybe add an explicit link to the example.

Other than that, I think this is good to go. @crvernon Let me know what you think.
