### 2022-03-09 #3709 @omshinde
@AlexanderJuestel @crvernon
I am ready with my final review and I have checked off the remaining items. The authors have done a great job to provide neat software and I congratulate them for their contributions. The updated documentation is comprehensive and detailed. Thank you to the authors for their efforts. From my side, I am happy to recommend the work to be published in JOSS.

Thanks, @crvernon for the follow-up and for the opportunity to review the manuscript.
