### 2022-05-02 #3709 @crvernon
@AlexanderJuestel  we are almost there!  Next is just setting up the archive for your new release.

- We want to make sure the archival has the correct metadata that JOSS requires.  This includes **a title that matches the paper title and a correct author list.**

So here is what we have left to do:
- [ ] Conduct a GitHub release of the current reviewed version of the software you now have on the `main` and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
- [ ] Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title) and author list. You may also add the authors' ORCID.
- [ ] Please list the DOI of the archived version here
