### 2022-05-02 #3709 @AlexanderJuestel
Hi @crvernon,

please find the archived version of GemGIS here under this DOI Number: https://doi.org/10.5281/zenodo.6511767
